Job list and search end-point in separate resource file
