+++ tapis-jobsapi/src/main/java/edu/utexas/tacc/tapis/jobs/api/resources/JobListingResource.java
@@ -23,13 +23,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import edu.utexas.tacc.tapis.jobs.api.responses.RespGetJobList;
-import edu.utexas.tacc.tapis.jobs.api.responses.RespJobSearch;
-import edu.utexas.tacc.tapis.jobs.api.responses.RespJobSearchAllAttributes;
 import edu.utexas.tacc.tapis.jobs.api.utils.JobsApiUtils;
 import edu.utexas.tacc.tapis.jobs.impl.JobsImpl;
-import edu.utexas.tacc.tapis.jobs.model.Job;
 import edu.utexas.tacc.tapis.jobs.model.dto.JobListDTO;
-import edu.utexas.tacc.tapis.search.SearchUtils;
 import edu.utexas.tacc.tapis.shared.exceptions.TapisImplException;
 import edu.utexas.tacc.tapis.shared.i18n.MsgUtils;
 import edu.utexas.tacc.tapis.shared.threadlocal.SearchParameters;
@@ -42,7 +38,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 
-	@Path("/")
+	@Path("/list")
 	public class JobListingResource 
 	 extends AbstractResource
 	{
@@ -101,7 +97,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 	     /*                                Public Methods                                */
 	     /* **************************************************************************** */
 	     /* ---------------------------------------------------------------------------- */
-	     /* getJobList:                                                                      */
+	     /* getJobList:                                                                  */
 	     /* ---------------------------------------------------------------------------- */
 	     @GET
 	     @Produces(MediaType.APPLICATION_JSON)
@@ -226,188 +222,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 	     }
 	     
 	     
-	     /* ---------------------------------------------------------------------------- */
-	     /* getJobSearchList:                                                                      */
-	     /* ---------------------------------------------------------------------------- */
-	     @GET
-	     @Path("/search")
-	     @Produces(MediaType.APPLICATION_JSON)
-	     @Operation(
-	             description = "Retrieve list of jobs for the user based on search conditions in the query paramter on the dedicated search end-point.\n\n"
-	                           + "The caller must be the job owner, creator or a tenant administrator."
-	                           + "",
-	             tags = "jobs",
-	             security = {@SecurityRequirement(name = "TapisJWT")},
-	             responses = 
-	                 {
-	                  @ApiResponse(responseCode = "200", description = "Jobs Search List retrieved.",
-	                      content = @Content(schema = @Schema(
-	                    		  implementation = edu.utexas.tacc.tapis.jobs.api.responses.RespJobSearch.class))),
-	                  @ApiResponse(responseCode = "400", description = "Input error.",
-	                      content = @Content(schema = @Schema(
-	                         implementation = edu.utexas.tacc.tapis.sharedapi.responses.RespBasic.class))),
-	                  @ApiResponse(responseCode = "401", description = "Not authorized.",
-	                      content = @Content(schema = @Schema(
-	                         implementation = edu.utexas.tacc.tapis.sharedapi.responses.RespBasic.class))),
-	                  @ApiResponse(responseCode = "403", description = "Forbidden.",
-	                      content = @Content(schema = @Schema(
-	                         implementation = edu.utexas.tacc.tapis.sharedapi.responses.RespBasic.class))),
-	                  @ApiResponse(responseCode = "404", description = "Jobs not found.",
-	                      content = @Content(schema = @Schema(
-	                         implementation = edu.utexas.tacc.tapis.sharedapi.responses.RespName.class))),
-	                  @ApiResponse(responseCode = "500", description = "Server error.",
-	                      content = @Content(schema = @Schema(
-	                         implementation = edu.utexas.tacc.tapis.sharedapi.responses.RespBasic.class)))}
-	     )
-	     public Response getJobSearchList(
-	    		 		@QueryParam("limit") int limit, 
-	    		 		@QueryParam("skip") int skip,
-	    		 		@QueryParam("startAfter") int startAfter,
-	    		 		@QueryParam("orderBy") String orderBy,
-	    		 		@QueryParam("computeTotal") boolean computeTotal,
-	    		 		@QueryParam("select") String select,
-	                    @DefaultValue("false") @QueryParam("pretty") boolean prettyPrint)
-	                               
-	     {
-	       // Trace this request.
-	       if (_log.isTraceEnabled()) {
-	         String msg = MsgUtils.getMsg("TAPIS_TRACE_REQUEST", getClass().getSimpleName(), "getJobSearchList", 
-	                                      "  " + _request.getRequestURL());
-	         _log.trace(msg);
-	       }
-	       
-	       // ------------------------- Create Context ---------------------------
-	       // Validate the threadlocal content here so no subsequent code on this request needs to.
-	       TapisThreadContext threadContext = TapisThreadLocal.tapisThreadContext.get();
-	       if (!threadContext.validate()) {
-	           var msg = MsgUtils.getMsg("TAPIS_INVALID_THREADLOCAL_VALUE", "validate");
-	           _log.error(msg);
-	           return Response.status(Status.INTERNAL_SERVER_ERROR).
-	                   entity(TapisRestUtils.createErrorResponse(msg, prettyPrint)).build();
-	       }
-	       
-	       List<String> searchList;
-	       int totalCount = -1;
-	       
-	       try
-	       {
-	         searchList = SearchUtils.buildListFromQueryParms(_uriInfo.getQueryParameters());
-	       }
-	       catch (Exception e)
-	       {
-	    	  String msg = MsgUtils.getMsg("JOBS_SEARCH_LIST_ERROR",threadContext.getJwtTenantId(),threadContext.getJwtUser(), threadContext.getOboTenantId(),threadContext.getOboUser(), e.getMessage());
-	         _log.error(msg, e);
-	         return Response.status(Status.BAD_REQUEST).entity(TapisRestUtils.createErrorResponse(msg,prettyPrint)).build();
-	       }
-	       if (searchList != null && !searchList.isEmpty()) _log.debug("Using searchList. First value = " + searchList.get(0));
-	       // ThreadContext designed to never return null for SearchParameters
-	       SearchParameters srchParms = threadContext.getSearchParameters();
-	        
-	       if(srchParms.getLimit() == null) {srchParms.setLimit(SearchParameters.DEFAULT_LIMIT);}
-	       List<String> selectList = srchParms.getSelectList();
-	       boolean allAttributesInResponse = false;
-	       if ((!selectList.isEmpty() ) && selectList.contains("allAttributes") ) {
-	    	   allAttributesInResponse = true;
-	    	   _log.debug("allAttributesInResponse is set to true");
-	       } ;
-	       computeTotal = srchParms.getComputeTotal();
-	       
-	       var jobsImpl = JobsImpl.getInstance();
-	       
-	       // summary attributes
-	       List<JobListDTO> jobSummaryList = null;
-	       List<Job> jobs = null; 
-	       
-	       // If we need the count and there was a limit then we need to make a call
-	       if (computeTotal && srchParms.getLimit() > 0)
-	       {
-	         
-					try {
-						totalCount = jobsImpl.getJobsSearchListCountByUsername(threadContext.getOboUser(), threadContext.getOboTenantId(), searchList,
-								   srchParms.getOrderByList());
-					} catch (TapisImplException e) {
-						_log.error(e.getMessage(), e);
-				           return Response.status(JobsApiUtils.toHttpStatus(e.condition)).
-				                   entity(TapisRestUtils.createErrorResponse(e.getMessage(), prettyPrint)).build();
-					}
-				
-		    }
-	       
-	       if(allAttributesInResponse == false && selectList.isEmpty() ) {
-		       try {
-		          
-		          
-		           jobSummaryList = jobsImpl.getJobSearchListByUsername(threadContext.getOboUser(), threadContext.getOboTenantId(), searchList,
-		        		   srchParms.getOrderByList(), srchParms.getLimit(),srchParms.getSkip());                       
-		       }
-		       catch (TapisImplException e) {
-		           _log.error(e.getMessage(), e);
-		           return Response.status(JobsApiUtils.toHttpStatus(e.condition)).
-		                   entity(TapisRestUtils.createErrorResponse(e.getMessage(), prettyPrint)).build();
-		       }
-		       catch (Exception e) {
-		           _log.error(e.getMessage(), e);
-		           return Response.status(Status.INTERNAL_SERVER_ERROR).
-		                   entity(TapisRestUtils.createErrorResponse(e.getMessage(), prettyPrint)).build();
-		       }
-		       
-		       if(jobSummaryList.isEmpty()) {
-	               String msg =  MsgUtils.getMsg("JOBS_SEARCH_NO_JOBS_FOUND", threadContext.getOboTenantId(),threadContext.getOboUser());
-	               RespJobSearch r = new RespJobSearch(jobSummaryList,srchParms.getLimit(),srchParms.getOrderBy(),srchParms.getSkip(),srchParms.getStartAfter(),-1);
-	               return Response.status(Status.OK).entity(TapisRestUtils.createSuccessResponse(msg,prettyPrint,r)).build(); 
-	            }
-	       
-	       
-		       if (computeTotal && srchParms.getLimit() <= 0) totalCount = jobSummaryList.size();
-		       RespJobSearch r = new RespJobSearch(jobSummaryList,srchParms.getLimit(),srchParms.getOrderBy(),srchParms.getSkip(),srchParms.getStartAfter(),totalCount);
-		       return Response.status(Status.OK).entity(TapisRestUtils.createSuccessResponse(
-		               MsgUtils.getMsg("JOBS_SEARCH_RESULT_LIST_RETRIEVED", threadContext.getOboUser(), threadContext.getOboTenantId()), prettyPrint, r)).build(); 
-	       
-	       } else {
-	    	   // select is provided by the user,
-	    	   // select all attributes in the query to db
-	    	   // then select the attributes that the user provides 
-	    	   try {
-			          
-			          
-		           jobs = jobsImpl.getJobSearchAllAttributesByUsername(threadContext.getOboUser(), threadContext.getOboTenantId(), searchList,
-		        		   srchParms.getOrderByList(), srchParms.getLimit(),srchParms.getSkip());                       
-		       }
-		       catch (TapisImplException e) {
-		           _log.error(e.getMessage(), e);
-		           return Response.status(JobsApiUtils.toHttpStatus(e.condition)).
-		                   entity(TapisRestUtils.createErrorResponse(e.getMessage(), prettyPrint)).build();
-		       }
-		       catch (Exception e) {
-		           _log.error(e.getMessage(), e);
-		           return Response.status(Status.INTERNAL_SERVER_ERROR).
-		                   entity(TapisRestUtils.createErrorResponse(e.getMessage(), prettyPrint)).build();
-		       }
-		       
-		       if(jobs.isEmpty()) {
-	               String msg =  MsgUtils.getMsg("JOBS_SEARCH_NO_JOBS_FOUND", threadContext.getOboTenantId(),threadContext.getOboUser());
-	               RespJobSearchAllAttributes r = new RespJobSearchAllAttributes(jobs,srchParms.getLimit(),srchParms.getOrderBy(),srchParms.getSkip(),srchParms.getStartAfter(),totalCount);
-	               return Response.status(Status.OK).entity(TapisRestUtils.createSuccessResponse(msg,prettyPrint,r)).build(); 
-	            }
-	       }
-
-	       if (computeTotal && srchParms.getLimit() <= 0) totalCount = jobs.size();
-	       
-	       // customize the response
-	       if(!selectList.isEmpty() && !selectList.contains("allAttributes") && !selectList.contains("summaryAttributes") ) {
-	    	  //TODO selectable fields
-	    	   
-	       }
-	       // ------------------------- Process Results --------------------------
-	       // Success.
-	       RespJobSearchAllAttributes r = new RespJobSearchAllAttributes (jobs,srchParms.getLimit(),srchParms.getOrderBy(),srchParms.getSkip(),srchParms.getStartAfter(),totalCount);
-	       
-	       
-	       return Response.status(Status.OK).entity(TapisRestUtils.createSuccessResponse(
-	               MsgUtils.getMsg("JOBS_SEARCH_RESULT_LIST_RETRIEVED", threadContext.getOboUser(), threadContext.getOboTenantId()), prettyPrint, r)).build();
-
-		  
-	     }
+	    
 	}
